commit 05: criacao de templates

+++ projeto1/app01/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class User(models.Model):
     nome = models.CharField('nome',max_length=30)
-    telefone = models.IntegerField('telefone')
+    telefone = models.BigIntegerField('telefone')
     email = models.CharField('email',max_length=30)
 
     def __str__(self):
